requirements.txt
opentelemetry-instrumentation-httpx
httpx
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
fastapi
uvicorn